Add a placeholder sounds module.
While writing this, I realized that I have to commit to either
supporting luajit or regular lua when I include either lua
implementation. I guess that would be hard to avoid. So I'm going with
luajit for now, but it would be nice if, in the future, I could make it
easy for users to choose either one. For example, I could include both
implementations and give users a choice when they compile things.