Allow play to be called multiple times.
An alternative design would be to restart the sound.

It's not obvoius to me that "continue playing" is better, but it fits my
intuition on what a play button does on something like a music player,
and it fits the behavior of love.audio, which is good for my long-term
goals with this project.