Add a stop method.
It took me a fair bit of work to figure out how the AudioQueue interface
wants this to happen. It seems (as far as I can make educated guesses)
that they want you to call AudioQueueStop only from the callback, and
that even then they expect you to keep pumping through the queues if
you'd like to enable restarting the stream (which I do).