Organize; reduce use of L in audio thread.
The goal for the audio thread is to have ZERO use of L, as it's not
thread-safe!

Sadly, this means we can't release a sound as soon as it's done playing.
My solution is to add a sounds.update function which is intended to be
called periodically to ensure sounds are released eventually. In
practice, I can image some users loading a finite number of sounds and
never caring too much about the memory, but I still want to support the
use case of people in memory-critical situations or theoretically
wanting to open and close many many sound files.