Reorganize dirs and add love compatability.
It's a bit tricky to run this from the love game engine because we can't
use C modules (well, we sort of can, but I chose to avoid it to make it
easier to use this as a library) and we don't want to own the run loop
within love. I decided that beatz itself ought to be love-aware, since
this is easier for the user and isn't really so dirty within this code.