Add a note callback that can influence playing.
This is a nice way for someone using this library to either respond
(song is read-only) to the track, or to respond and influence (song is
read/write) playback. The influence supported is either stopping the
song at a certain point basically from a note-triggered callback, or to
choose to insert delays within playback which may be useful for game
interactions.