Add a Track class.
This structure is messy. There's a track class, a load_env, and a data
table that are all basically the same thing. Then I've got a main_track
which is owned by the data/track class, but not well distinguished.

I'll add an issue to clean this up a bit.